+++ src/nota_asr_server/__init__.py
@@ -1,3 +1,3 @@
 """Nota ASR Server."""
 
-__version__ = "0.3.0"
+__version__ = "0.4.0"